# feat: add configurable compression level for zip downloads (Folder Share page, downloadPageFolder.html)