Added the ability to customize zip file compression levels:
    
    - Added new ArozZipFileWithCompressionLevel function that accepts compression level parameter
    - Added compression_level query parameter support to share downloads
    - Added UI checkbox to toggle between compressed/uncompressed downloads
    - Auto-detect when folder contains mostly media files (>= 50% in size) and pre-check uncompressed option
    - Updated compression help text to provide guidance on media files
    - Default compression level remains at flate.DefaultCompression (-1)

The changes allow users to choose no compression for media-heavy folders to improve zipping speed, while maintaining regular compression for other content types.

![Screenshot From 2025-01-14 18-13-40](https://github.com/user-attachments/assets/eefe1f58-1c3b-4b8a-907f-cf13c247be70)

# Case for 500 jpegs of 2.5GB in total
**10-20x faster processing if uncompressed**

Before (default compression zipping)
```
[saren@saren-gaming ~ ]% time curl 'https://AROZ/share/download/f5b58501-b363-42e1-8ae9-2c48151b8a44' > /dev/null 
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100 2554M  100 2554M    0     0  58.9M      0  0:00:43  0:00:43 --:--:--  626M
curl  > /dev/null  0.53s user 1.68s system 5% cpu 43.322 total
```

After (uncompressed zipping)
```
[saren@saren-gaming ~ ]% time curl 'https://AROZ/share/download/f5b58501-b363-42e1-8ae9-2c48151b8a44?compression_level=0' > /dev/null
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100 2564M  100 2564M    0     0   452M      0  0:00:05  0:00:05 --:--:--  575M
curl  > /dev/null  0.58s user 1.48s system 36% cpu 5.668 total
```